+++ src/qudi/core/gui/main_gui/__init__.py
@@ -1,3 +1,0 @@
-# -*- coding: utf-8 -*-
-
-from qudi.core.gui.main_gui.main_gui import QudiMainGui